# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Coin Change API
  description: API for calculating minimum coins for a given amount
paths:
  /api/v1/coin-change/minimum-coins:
    post:
      tags:
      - Coin Change API
      summary: Calculate minimum coins
      description: Calculate the minimum number of coins needed to make up the given
        amount
      operationId: getMinimumCoins
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoinChangeRequest"
        required: true
      responses:
        "400":
          description: Invalid input provided
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Successfully calculated coins
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CoinChangeResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CoinChangeResponse"
components:
  schemas:
    CoinChangeRequest:
      required:
      - amount
      - denominations
      type: object
      properties:
        amount:
          maximum: 10000.00
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
          description: The amount for which to calculate minimum coins
          example: 123.45
        denominations:
          type: array
          description: List of coin denominations
          example:
          - 1
          - 5
          - 10
          - 25
          items:
            type: number
            description: List of coin denominations
    CoinChangeResponse:
      type: object
      properties:
        coins:
          type: array
          description: List of coins in ascending order
          example:
          - 1
          - 5
          - 10
          - 25
          items:
            type: number
            description: List of coins in ascending order